Wait for 4DKeyBinds.dll
Wait for 4DKeyBinds.dll to get loaded before calling its functions. Timeout for good after 2s.

If someone doesn't like this they can remove it, but it's good to have defaults that work reliably. Without this, mods that are sorted alphabetically before 4DKeyBinds.dll will not be able to add binds.